research/requirements.txt
polars_ta>=0.2.3
expr_codegen>=0.6.0
alphainspect>=0.3.0